+++ docs/changelog.rst
@@ -42,6 +42,7 @@ v1.0.0 (in development)
 - Renamed ``DistributionPackage.from_pep691_details()`` to
   `~DistributionPackage.from_json_data()`
 - `PyPISimple.stream_project_names()` now accepts JSON responses
+- Use pydantic internally to parse JSON responses
 
 v0.10.0 (2022-06-30)
 --------------------